chore(1) - functionality for creating custom button
