+++ Bkit/Sources/Bkit/Theme/ButtonStyle+Enum.swift
@@ -1,0 +1,8 @@
+
+import Foundation
+
+public enum ButtonStyle {
+    case rounded
+    case rectangle
+    case buttonRounded
+}
