+++ Bkit/Sources/Bkit/Theme/Size.swift
@@ -1,0 +1,12 @@
+
+import Foundation
+
+public struct Size {
+    let width: CGFloat
+    let height: CGFloat
+    
+    public init(width: CGFloat, height: CGFloat) {
+        self.width = width
+        self.height = height
+    }
+}
